refactor: library.api 패키지를 library.reader 하위 패키지로 이동
